build: modify the setup.cfg file
In preparation for the upload on PyPi, the setup.cfg file has been
modified to include better classification of the project and to update
the version of the project.

Since the project is in fact currently used in production and it
proved stable enough, it's version has been upgraded to 1.0.0 and it's
status has been upgraded from development to production/stable.