feat: implement a new logic for the download manifest writer
This commit upgrades the download manifest writer to tackle those
situations where the intial download of the files was missing one or
more files, and the missing files are downloaded at a later point in
time, once uploaded in the data platform.

In the previous implementation, the above situation would result in
the original download manifest file being overwritten, resulting in
only the data of the latest download being in the file.

The new implementation corrects this behaviour, and the download
manifest writer now checks whether an existing manifest file
is already present at the indicated path and if a file is found,
the program now updates the existing manifest file with the new
information. This new implementation preserves the original data
in the manifest file, and update the list of downloaded files
with the newly discovered files' data.