fix: modify the download manifest update process
In the previous implementation, the update procedure of the in memory
download manifest consisted in copying the content of the existing
download manifest (a list of dictionaries) and extending it with the
new download details. However, since the extend method does not
return an object and modifies the original list in place, the
update process was resulting in a None object being written to the
file.

This implementation modifies the previous one by now separating the
copying step and the update step, such that the proper object (an
updated list of download details) is produced at the end of the
update process.