fix: modify the handling of existing entries during manifest update
Before this fix was introduced, if the update of the manifest file
interested some files that were already present in the original
manifest file, at the end of the update process, the already
existing entry would be included twice in the updated manifest
file.

To correct this behaviour, we have now included a further check
in the updating process to guarantee that only new file instances
are included in the updated manifest.The update_manifest_file
function now includes in the updated manifest file only those files
that were not already present in the original manifest file, thus
avoiding the presence of duplicates in the updated manifest file.